Do not recurse into child nodes in find Iterator
The find iterator recursively iterated over it's left/right children's children
before adding them to the queue. I believe this is not necessary and causes nodes
to be vsisted multiple times. This patch only uses the candidate's node max
and start value to decide if further traversal is needed.

I observe about a 2 to 4-fold performance gain.