Remove some uses of unsafe from Myers pattern matching and fastexp
There's two uses of `unsafe` that aren't necessary:

- using a union to convert between raw bits in a f64 floating point
  format and f64 itself, because there's the safe f64::from_bits
  function in std

- using from_raw_parts to build a slice &[T] from a single &T, because
  there's the safe slice::from_ref function in std

The latter flags a minor safety issue: the
ShortTracebackHandler::column_slice method had an incorrect lifetime
in its return value. It was incorrectly suggesting that the returned
slice lived for longer than the `self` value. In particular, code
could write something like (in pseudo-Rust, with named lifetimes of
locals):

    let states: &'outer [State<T, T::DistType>] = ...;
    let column: &'outer [State<T, T::DistType>];
    {
        let handler = ShortTracebackHandler::new(.., states);
        column = handler.column_slice(); // refers to handler.state
    } // handler becomes invalid, as does handler.state

    println!("{}", column); // invalid use of reference to handler.state

This method is private and explicitly designed for debugging use only,
and so (a) this minor safety issue probably doesn't cause issues in
practice, and (b) the fix doesn't break any existing code.