feat: Add to_string method for Newick IO
This adds the `to_string`, `to_file`, and `write` methods to
`newick.rs`, mirroring the existing `from_string`, `from_file` and
`read` methods.

A slight annoyance is that petgraph iterates outgoing edges in reverse
order of adding them. To keep input and output consistent, I manually
reverse that order back to the natural order in which they were added.
Another option would be to actually add them in the reverse order, but
that would change the representation which may not be desirable.

I'm also changing the `"N/A"` value for nodes without a name to just the
empty string as this heavily pollutes the `to_string` representation,
and `""` is more naturaly IMO.

Also fixed one unused variable warning.