feat!: Update `strum` and `ordered-float` dependencies and change From<LogProb> into TryFrom<LogProb> for NotNan<f64>. (#491)
* Update `strum` crates

Figure we'll need to change eventually.

* Upgrade Ordered Float

ordered-float ditched their From<f64> methods for NotNan and instead made them TryFrom<f64> (https://github.com/reem/rust-ordered-float/commit/39f76bb4910eeb4ad5c43f4111fefc550c258460).  This change makes sense because From trait implementations generally shouldn’t panic, which the implementation did if the f64 was a NaN, and so TryFrom is more appropriate.  This commit accounts for this change in the upgrade

* Remove Deprecated Function Call

`ndarray` has renamed `genrows()` to `rows()`

* Formatting Fix

* Update CHANGELOG

* `From -> `TryFrom`

The conversion from LogProb is not guaranteed to succeed, so we use `TryFrom` instead of `From`, as per the Rust docs:

```Note: The From trait must not fail. The From trait is intended for perfect conversions. If the conversion can fail or is not perfect, use TryFrom.```

Note this does change the API.

* Apply suggestions from code review

Co-authored-by: Adam Azarchs <adam.azarchs@10xgenomics.com>

* Update CHANGELOG.md

* Update Cargo.toml

* Fix build issue.

Co-authored-by: Adam Azarchs <adam.azarchs@10xgenomics.com>
Co-authored-by: Johannes Köster <johannes.koester@uni-due.de>