Add to_bytes and from_bytes for Occ
to_bytes creates a byte vector from Occ struct, while from_bytes
reverses the process. This makes it easier for users who want to create
a set of index files for alignment purposes. For large genomes, creating
the Occ struct takes a non-negligible amount of time, so it is
preferable to do this in a single step prior to future alignments. Prior
to this commit, the internals of Occ weren't available in any capacity.
While the internals are still unavailable to users (from the Occ struct
itself), they are able to save the contents to a file and recreate it in
downstream steps.