Improve MyersBuilder::ambig() method (small behaviour change)
Adds the possibility to call ambig() multiple times
on the same symbol without overwriting the
previous equivalent list.
This may be more intuitive and simplifies some code
(e.g. for case-insensitive matching).
But it is probably a breaking change, since the behaviour
is different, although it is very unlikely to break anything
existing.

Also removed `.chain(Some(byte))`, since this was
unnecessary.